# Add recommended inductor config for speedup