Tested on A100.
```bash
w/o recommended_inductor_config_setter()
68%|███████████████████▊         | 684/999 [00:05<00:02, 125.37it/s]
w/ recommended_inductor_config_setter()
67%|█████████████████████████▍            | 669/999 [00:04<00:02, 144.30it/s]




w/ recommended_inductor_config_setter()
----------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
                                                   Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg    # of Calls  
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
void at::native::tinygemm_m16n8k16_chunk_kernel<at::...         0.00%       0.000us         0.00%       0.000us       0.000us        3.016s        74.83%        3.016s      18.444us        163520  
                                 triton_red_fused_bmm_4         0.00%       0.000us         0.00%       0.000us       0.000us     296.429ms         7.36%     296.429ms      12.690us         23360  
                                 triton_red_fused_bmm_2         0.00%       0.000us         0.00%       0.000us       0.000us     158.435ms         3.93%     158.435ms       6.782us         23360  
                             triton_red_fused_div_mm_16         0.00%       0.000us         0.00%       0.000us       0.000us      98.043ms         2.43%      98.043ms     134.306us           730  
                        triton_per_fused__softmax_add_3         0.00%       0.000us         0.00%       0.000us       0.000us      56.517ms         1.40%      56.517ms       2.419us         23360  
     triton_poi_fused__to_copy_add_cat_index_copy_mul_1         0.00%       0.000us         0.00%       0.000us       0.000us      52.315ms         1.30%      52.315ms       2.240us         23360  
                        triton_per_fused__to_copy_bmm_5         0.00%       0.000us         0.00%       0.000us       0.000us      51.179ms         1.27%      51.179ms       2.191us         23360  
                            triton_poi_fused_mul_silu_7         0.00%       0.000us         0.00%       0.000us       0.000us      46.271ms         1.15%      46.271ms       1.981us         23360  
    triton_red_fused__to_copy_add_mean_mul_pow_rsqrt_12         0.00%       0.000us         0.00%       0.000us       0.000us      39.436ms         0.98%      39.436ms       3.601us         10950  
    triton_red_fused__to_copy_add_mean_mul_pow_rsqrt_10         0.00%       0.000us         0.00%       0.000us       0.000us      37.628ms         0.93%      37.628ms       3.436us         10950  
    triton_red_fused__to_copy_add_mean_mul_pow_rsqrt_11         0.00%       0.000us         0.00%       0.000us       0.000us      35.661ms         0.88%      35.661ms       3.257us         10950  
    triton_red_fused__to_copy_add_mean_mul_pow_rsqrt_14         0.00%       0.000us         0.00%       0.000us       0.000us      35.401ms         0.88%      35.401ms       3.464us         10220  
triton_red_fused__softmax__to_copy_argmax_div_expone...         0.00%       0.000us         0.00%       0.000us       0.000us      24.421ms         0.61%      24.421ms      33.453us           730  
void at::native::mbtopk::radixFindKthValues<float, u...         0.00%       0.000us         0.00%       0.000us       0.000us      15.414ms         0.38%      15.414ms       5.279us          2920  
void at::native::mbtopk::computeBlockwiseWithinKCoun...         0.00%       0.000us         0.00%       0.000us       0.000us       6.487ms         0.16%       6.487ms       2.222us          2920  
void at_cuda_detail::cub::DeviceScanByKeyKernel<at_c...         0.00%       0.000us         0.00%       0.000us       0.000us       5.331ms         0.13%       5.331ms       3.652us          1460  
void at::native::vectorized_elementwise_kernel<4, at...         0.00%       0.000us         0.00%       0.000us       0.000us       4.434ms         0.11%       4.434ms       2.026us          2189  
void at::native::mbtopk::gatherTopK<float, unsigned ...         0.00%       0.000us         0.00%       0.000us       0.000us       3.969ms         0.10%       3.969ms       5.436us           730  
                         Memcpy DtoD (Device -> Device)         0.00%       0.000us         0.00%       0.000us       0.000us       3.948ms         0.10%       3.948ms       1.804us          2188  
void at::native::bitonicSortKVInPlace<-2, -1, 16, 16...         0.00%       0.000us         0.00%       0.000us       0.000us       3.933ms         0.10%       3.933ms       5.388us           730  
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
Self CPU time total: 4.395s
Self CUDA time total: 4.030s

w/o recommended_inductor_config_setter()
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
                                                   Name    Self CPU %      Self CPU   CPU total %     CPU total  CPU time avg     Self CUDA   Self CUDA %    CUDA total  CUDA time avg    # of Calls  
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
void at::native::tinygemm_m16n8k16_chunk_kernel<at::...         0.00%       0.000us         0.00%       0.000us       0.000us        3.509s        65.84%        3.509s      19.246us        182336  
                                                triton_         0.00%       0.000us         0.00%       0.000us       0.000us        1.086s        20.38%        1.086s       5.035us        215710  
std::enable_if<!(false), void>::type internal::gemvx...         0.00%       0.000us         0.00%       0.000us       0.000us     426.334ms         8.00%     426.334ms       8.184us         52096  
ampere_bf16_s16816gemm_bf16_128x64_ldg8_f2f_stages_6...         0.00%       0.000us         0.00%       0.000us       0.000us     123.559ms         2.32%     123.559ms     151.793us           814  
                                          memcpy32_post         0.00%       0.000us         0.00%       0.000us       0.000us     107.750ms         2.02%     107.750ms       2.036us         52910  
void at::native::mbtopk::radixFindKthValues<float, u...         0.00%       0.000us         0.00%       0.000us       0.000us      17.418ms         0.33%      17.418ms       5.350us          3256  
void at::native::mbtopk::computeBlockwiseWithinKCoun...         0.00%       0.000us         0.00%       0.000us       0.000us       7.294ms         0.14%       7.294ms       2.240us          3256  
void at_cuda_detail::cub::DeviceScanByKeyKernel<at_c...         0.00%       0.000us         0.00%       0.000us       0.000us       5.936ms         0.11%       5.936ms       3.646us          1628  
void at::native::vectorized_elementwise_kernel<4, at...         0.00%       0.000us         0.00%       0.000us       0.000us       4.938ms         0.09%       4.938ms       2.023us          2441  
void cublasLt::splitKreduce_kernel<32, 16, int, __nv...         0.00%       0.000us         0.00%       0.000us       0.000us       4.895ms         0.09%       4.895ms       6.013us           814  
void at::native::bitonicSortKVInPlace<-2, -1, 16, 16...         0.00%       0.000us         0.00%       0.000us       0.000us       4.683ms         0.09%       4.683ms       5.753us           814  
void at::native::mbtopk::gatherTopK<float, unsigned ...         0.00%       0.000us         0.00%       0.000us       0.000us       4.644ms         0.09%       4.644ms       5.705us           814  
                         Memcpy DtoD (Device -> Device)         0.00%       0.000us         0.00%       0.000us       0.000us       4.522ms         0.08%       4.522ms       1.853us          2440  
                         Memcpy DtoH (Device -> Pinned)         0.00%       0.000us         0.00%       0.000us       0.000us       4.030ms         0.08%       4.030ms       2.475us          1628  
void at::native::index_elementwise_kernel<128, 4, at...         0.00%       0.000us         0.00%       0.000us       0.000us       3.560ms         0.07%       3.560ms       4.373us           814  
void at::native::vectorized_elementwise_kernel<4, at...         0.00%       0.000us         0.00%       0.000us       0.000us       2.650ms         0.05%       2.650ms       1.628us          1628  
void at_cuda_detail::cub::DeviceScanByKeyInitKernel<...         0.00%       0.000us         0.00%       0.000us       0.000us       2.070ms         0.04%       2.070ms       1.271us          1628  
                       Memcpy DtoH (Device -> Pageable)         0.00%       0.000us         0.00%       0.000us       0.000us       1.945ms         0.04%       1.945ms       2.387us           815  
void at::native::(anonymous namespace)::distribution...         0.00%       0.000us         0.00%       0.000us       0.000us       1.941ms         0.04%       1.941ms       2.384us           814  
void gemmk1_kernel<int, float, 256, 5, true, false, ...         0.00%       0.000us         0.00%       0.000us       0.000us       1.778ms         0.03%       1.778ms       2.184us           814  
-------------------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  ------------  
Self CPU time total: 5.949s
Self CUDA time total: 5.330s

```